Refactor file paths and import statements, add Server class to services/index.js, remove constants/messages/index.js, add constants/index.js, update routes and controllers imports
