+++ app.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
 
-const Server = require('./src/services/server.service');
+const { Server } = require('./src/services');
 
 const server = new Server();
