+++ constants/index.js
@@ -1,0 +1,7 @@
+const api = require('./api/api.url');
+const httpResponses = require('./messages/http-responses');
+
+module.exports = {
+    api,
+    ...httpResponses,
+}
