--- constants/messages/index.js
@@ -1,5 +1,0 @@
-const httpResponses = require('./http-responses');
-
-module.exports = {
-    ...httpResponses,    
-}
