+++ src/controllers/index.js
@@ -1,9 +1,9 @@
 const auth = require("./auth.controller");
-const category = require("./category.controller");
-const product = require("./product.controller");
+const category = require("./categories.controller");
+const product = require("./products.controller");
 const search = require("./search.controller");
-const upload = require("./upload.controller");
-const user = require("./user.controller");
+const upload = require("./uploads.controller");
+const user = require("./users.controller");
 
 module.exports = {
     ...auth,
