+++ src/middlewares/validate-file.js
@@ -1,5 +1,5 @@
 const { response } = require("express");
-const { fileNotExist } = require("../../constants/messages");
+const { fileNotExist } = require("../../constants");
 
 const validateFileToBeUpload = (req, res = response, next) => {
   if (!req.files || Object.keys(req.files).length === 0 || !req.files.file) {
