+++ src/middlewares/validate-jwt.middleware.js
@@ -1,7 +1,7 @@
 const { response } = require("express");
 const jwt = require("jsonwebtoken");
 const User = require("../models/user.model");
-const { httpResponses } = require("../../constants/messages/http-responses");
+const { httpResponses } = require("../../constants");
 
 const { status401 } = httpResponses;
