+++ src/models/index.js
@@ -1,11 +1,9 @@
-const Server = require('../services/server.service');
 const Category = require('./category.model');
 const Product = require('./product.model');
 const Role = require('./role.model');
 const User = require('./user.model');
 
-module.exports = {
-    Server,
+module.exports = {    
     Category,
     Product,
     Role,
