+++ src/routes/uploads.routes.js
@@ -1,6 +1,6 @@
 const { Router } = require("express");
 const { check } = require("express-validator");
-const { uploads, updateUpload, updatePictureCloudinary, showImages } = require("../controllers/uploads.controller");
+const { uploads, updateUpload, updatePictureCloudinary, showImages } = require("../controllers");
 const { allowedCollections } = require("../../helpers");
 const { validateFields, validateFileToBeUpload } = require("../middlewares");
