+++ src/services/index.js
@@ -1,0 +1,5 @@
+const Server = require('./server.service');
+
+module.exports = {
+    Server    
+}
